+++ sampleproject/Assets/Samples/Asteroids/Client/Systems/LoadLevelSystem.cs
@@ -8,25 +8,28 @@ namespace Asteroids.Client
     public class LoadLevelSystem : SystemBase
     {
         private BeginSimulationEntityCommandBufferSystem m_Barrier;
-        private RpcQueue<RpcLevelLoaded> m_RpcQueue;
-        private EntityQuery m_LevelGroup;
+        private RpcQueue<RpcLevelLoaded, RpcLevelLoaded> m_RpcQueue;
         private Entity m_LevelSingleton;
 
         protected override void OnCreate()
         {
             m_Barrier = World.GetOrCreateSystem<BeginSimulationEntityCommandBufferSystem>();
-            m_RpcQueue = World.GetOrCreateSystem<RpcSystem>().GetRpcQueue<RpcLevelLoaded>();
+            m_RpcQueue = World.GetOrCreateSystem<RpcSystem>().GetRpcQueue<RpcLevelLoaded, RpcLevelLoaded>();
 
-            // The level always exist, "loading" just resizes it
-            m_LevelSingleton = EntityManager.CreateEntity();
-            EntityManager.AddComponentData(m_LevelSingleton, new LevelComponent {width = 0, height = 0});
-            m_LevelGroup = GetEntityQuery(ComponentType.ReadWrite<LevelComponent>());
-            RequireForUpdate(m_LevelGroup);
+            RequireForUpdate(GetEntityQuery(ComponentType.ReadOnly<LevelLoadRequest>(), ComponentType.ReadOnly<ReceiveRpcCommandRequestComponent>()));
+            // This is just here to make sure the subscen is streamed in before the client sets up the level data
+            RequireSingletonForUpdate<GhostPrefabCollectionComponent>();
         }
 
         protected override void OnUpdate()
         {
-            var commandBuffer = m_Barrier.CreateCommandBuffer().ToConcurrent();
+            if (!HasSingleton<LevelComponent>())
+            {
+                // The level always exist, "loading" just resizes it
+                m_LevelSingleton = EntityManager.CreateEntity();
+                EntityManager.AddComponentData(m_LevelSingleton, new LevelComponent {width = 0, height = 0});
+            }
+            var commandBuffer = m_Barrier.CreateCommandBuffer().AsParallelWriter();
             var rpcFromEntity = GetBufferFromEntity<OutgoingRpcDataStreamBufferComponent>();
             var levelFromEntity = GetComponentDataFromEntity<LevelComponent>();
             var levelSingleton = m_LevelSingleton;
@@ -35,7 +38,7 @@ namespace Asteroids.Client
             {
                 commandBuffer.DestroyEntity(nativeThreadIndex, entity);
                 // Check for disconnects
-                if (!rpcFromEntity.Exists(requestSource.SourceConnection))
+                if (!rpcFromEntity.HasComponent(requestSource.SourceConnection))
                     return;
                 // set the level size - fake loading of level
                 levelFromEntity[levelSingleton] = new LevelComponent