+++ sampleproject/Assets/Samples/Asteroids/Client/Systems/Rendering/ShipRenderSystem.cs
@@ -8,7 +8,6 @@ using Unity.NetCode;
 
 namespace Asteroids.Client
 {
-    [UpdateAfter(typeof(RenderInterpolationSystem))]
     [UpdateBefore(typeof(ParticleEmitterSystem))]
     [UpdateInGroup(typeof(ClientPresentationSystemGroup))]
     public class ShipThrustParticleSystem : SystemBase
@@ -22,7 +21,6 @@ namespace Asteroids.Client
         }
     }
 
-    [UpdateAfter(typeof(RenderInterpolationSystem))]
     [UpdateBefore(typeof(LineRenderSystem))]
     [UpdateInGroup(typeof(ClientPresentationSystemGroup))]
     public class ShipTrackingSystem : SystemBase
@@ -62,14 +60,14 @@ namespace Asteroids.Client
             var level = m_LevelGroup.ToComponentDataArrayAsync<LevelComponent>(Allocator.TempJob, out levelHandle);
             var teleport = m_Teleport;
 
-            var trackJob = Entities.WithReadOnly(shipPosition).WithReadOnly(level).WithDeallocateOnJobCompletion(level).
+            var trackJob = Entities.WithReadOnly(shipPosition).WithReadOnly(level).WithDisposeOnCompletion(level).
                 ForEach((ref LineRendererComponentData target) =>
             {
                 int mapWidth = level[0].width;
                 int mapHeight = level[0].height;
                 int nextTeleport = 1;
 
-                if (shipPosition.Exists(localPlayerShip))
+                if (shipPosition.HasComponent(localPlayerShip))
                 {
                     float3 pos = shipPosition[localPlayerShip].Value;
                     pos.x -= screenWidth / 2;
@@ -94,7 +92,6 @@ namespace Asteroids.Client
     }
 
     [UpdateBefore(typeof(LineRenderSystem))]
-    [UpdateAfter(typeof(RenderInterpolationSystem))]
     [UpdateInGroup(typeof(ClientPresentationSystemGroup))]
     public class ShipRenderSystem : SystemBase
     {