+++ sampleproject/Assets/Samples/Asteroids/GameMain.cs
@@ -14,7 +14,7 @@ using Unity.NetCode.Editor;
 [UpdateBefore(typeof(TickClientSimulationSystem))]
 #endif
 [UpdateInWorld(UpdateInWorld.TargetWorld.Default)]
-public class AsteroidsClientServerControlSystem : ComponentSystem
+public class AsteroidsClientServerControlSystem : SystemBase
 {
     private const ushort networkPort = 50001;
 
@@ -40,10 +40,6 @@ public class AsteroidsClientServerControlSystem : ComponentSystem
             if (world.GetExistingSystem<ServerSimulationSystemGroup>() != null)
             {
                 var entityManager = world.EntityManager;
-                var settings = entityManager.CreateEntity();
-                var settingsData = GetSingleton<ServerSettings>();
-                entityManager.AddComponentData(settings, settingsData);
-
                 var grid = entityManager.CreateEntity();
                 entityManager.AddComponentData(grid, new GhostDistanceImportance
                 {
@@ -61,8 +57,6 @@ public class AsteroidsClientServerControlSystem : ComponentSystem
             // Auto connect all clients to the server
             if (world.GetExistingSystem<ClientSimulationSystemGroup>() != null)
             {
-                // Enable fixed tick rate
-                world.EntityManager.CreateEntity(typeof(FixedClientTickRate));
                 NetworkEndPoint ep = NetworkEndPoint.LoopbackIpv4;
                 ep.Port = networkPort;
 #if UNITY_EDITOR
@@ -84,7 +78,9 @@ public class GameMain : UnityEngine.MonoBehaviour, IConvertGameObjectToEntity
     public int numAsteroids = 200;
     public int levelWidth = 2048;
     public int levelHeight = 2048;
-    public int damageShips = 1;
+    public bool damageShips = true;
+    public int relevancyRadius = 0;
+    public bool staticAsteroidOptimization = false;
     public void Convert(Entity entity, EntityManager dstManager, GameObjectConversionSystem conversionSystem)
     {
 #if !UNITY_CLIENT || UNITY_SERVER || UNITY_EDITOR
@@ -98,6 +94,8 @@ public class GameMain : UnityEngine.MonoBehaviour, IConvertGameObjectToEntity
         settings.levelWidth = levelWidth;
         settings.levelHeight = levelHeight;
         settings.damageShips = damageShips;
+        settings.relevancyRadius = relevancyRadius;
+        settings.staticAsteroidOptimization = staticAsteroidOptimization;
         dstManager.AddComponentData(entity, settings);
 #endif
     }